chore: MainTabBarCoordinator 삭제
