+++ Projects/App/Sources/AppCoordinator.swift
@@ -1,11 +1,17 @@
 import UIKit
 
+import AgencyFeature
+import AgencyFeatureInterface
 import BaseFeature
 import BaseFeatureInterface
 import MainFeature
 import SignFeature
 import SignFeatureInterface
 import DesignSystem
+import LedgerFeature
+import LedgerFeatureInterface
+import MyPageFeature
+import MyPageFeatureInterface
 
 final class AppCoordinator: Coordinator {
   var navigationController: UINavigationController
@@ -30,12 +36,12 @@ final class AppCoordinator: Coordinator {
       sign(animated: true)
     case .ledger:
       main(animated: true)
-      let mainCoordinator = childCoordinators.first { $0 is MainTabBarCoordinator }
-      mainCoordinator?.move(to: .ledger)
+//      let mainCoordinator = childCoordinators.first { $0 is MainTabBarCoordinator }
+//      mainCoordinator?.move(to: .ledger)
     case let .createManualLedger(id):
       main(animated: true)
-      let mainCoordinator = childCoordinators.first { $0 is MainTabBarCoordinator }
-      mainCoordinator?.move(to: .createManualLedger(id))
+//      let mainCoordinator = childCoordinators.first { $0 is MainTabBarCoordinator }
+//      mainCoordinator?.move(to: .createManualLedger(id))
     default: break
     }
   }
@@ -53,11 +59,35 @@ extension AppCoordinator {
   }
   
   func main(animated: Bool) {
-    let mainTabCoordinator = MainTabBarCoordinator(
-      navigationController: navigationController
+    let mainTapVC = MainTapViewController()
+    mainTapVC.setViewControllers(
+      [
+        agencyTab(),
+        ledgerTab(),
+        myPageTab()
+      ],
+      animated: false
     )
-    mainTabCoordinator.start(animated: true)
-    mainTabCoordinator.parentCoordinator = self
-    childCoordinators.append(mainTabCoordinator)
+    navigationController.isNavigationBarHidden = true
+    navigationController.viewControllers = [mainTapVC]
+  }
+  
+  private func agencyTab() -> UIViewController {
+    let factory = DIContainer.shared.resolve(type: AgencyFactoryInterface.self)
+    let agencyListVC = factory.makeAgencyList()
+    return UINavigationController(rootViewController: agencyListVC)
+  }
+  
+  private func ledgerTab() -> UIViewController {
+    let ledgerFactory = DIContainer.shared.resolve(type: LedgerFactoryInterface.self)
+    let ledgerTab = ledgerFactory.makeLedgerTab()
+    let memberTab = ledgerFactory.makeMemberTab(delegate: nil)
+    let ledgerMainVC = ledgerFactory.makeLedgerMain(ledgerTab: ledgerTab, memberTab: memberTab)
+    return UINavigationController(rootViewController: ledgerMainVC)
+  }
+  
+  private func myPageTab() -> UIViewController {
+    let myPageVC = DIContainer.shared.resolve(type: MyPageFactoryInterface.self).makeMyPageVC()
+    return UINavigationController(rootViewController: myPageVC)
   }
 }
